Moving the decorator out of fileRoutes and into its own file
